+++ ChangeLog.md
@@ -14,6 +14,7 @@
     - Remove TTP::nodesDirs() function
     - Remove TTP::Node->dirs() method in favor of updated TTP::Node->finder() (todo #106)
     - Obsolete TTP::host() in favor of already existing TTP::nodeName() (todo #102)
+    - Obsolete 'credentialsDirs' variable in favor of 'credentials.dirs'
 
 ### 4.6.0